+++ src/test/static/namespaces/namespace-unnamed.js
@@ -1,0 +1,11 @@
+/**
+ * @namespace
+ */
+var ImplicitlyNamedNamespace = {};
+
+/**
+ * @namespace
+ */
+ImplicitlyNamedNamespace.NestedNamespace = {
+  foo: 'bar'
+};